the convert patch
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@9538 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/Spacing.h
@@ -35,7 +35,7 @@ public:
 	///
 	Spacing() : space(Default), value("1.0") {}
 	///
-	Spacing(Spacing::Space sp, float val = 1.0) {
+	Spacing(Spacing::Space sp, double val = 1.0) {
 		set(sp, val);
 	}
 	Spacing(Spacing::Space sp, std::string const & val) {
@@ -52,7 +52,7 @@ public:
 	///
 	Spacing::Space getSpace() const { return space; }
 	///
-	void set(Spacing::Space sp, float val = 1.0);
+	void set(Spacing::Space sp, double val = 1.0);
 	///
 	void set(Spacing::Space sp, std::string const & val);
 	///
